+++ src/components/child/ChildCard.tsx
@@ -1,18 +1,15 @@
 "use client";
 import React, { useState } from "react";
-import browserClient, { DEFAULT_PROFILE_IMAGE_URL } from "@/utils/supabase/client";
-import Link from "next/link";
+import browserClient from "@/utils/supabase/client";
 import Schedule from "./Schedule";
-import CakeIcon from "../../../public/child/cake-icon.svg";
-import InjectorIcon from "../../../public/child/injector-icon.svg";
-import RightArrowIcon from "../../../public/child/right-arrow-icon.svg";
 import { useVaccineQuery, useVaccineRecordQuery } from "@/query/useVaccineRecordQuery";
 import { ToastDescription } from "@radix-ui/react-toast";
 import { useToast } from "@/hooks/use-toast";
 import { Child } from "@/types/childType";
-import Image from "next/image";
+import ChildBasicInfo from "./ChildCardBasicInfo";
+import ChildCardVaccinationInfo from "./ChildCardVaccinationInfo";
 
-interface ChildCardProps {
+type ChildCardProps = {
   child?: Child; // 등록된 child가 없으면 undefined일 수 있음
   onEdit?: (child: Child) => void; // 수정 기능을 위해 child를 전달
   onDelete?: (childId: string) => void;
@@ -110,111 +107,17 @@ export const ChildCard = ({ child, onDelete }: ChildCardProps) => {
         {/* 아이 기본 정보 & 우리 아이 접종 내역 컨테이너 */}
         <div className="flex flex-col md:flex-row justify-start items-start w-full gap-6">
           {/* 아이 기본 정보 컨테이너 */}
-          <div className="w-full md:w-[50%] flex flex-col justify-start items-start flex-grow-0 flex-shrink-0 p-4 rounded-2xl bg-white shadow-[0px_0px_12px_#7272721A]">
-            <div className="w-full flex flex-col gap-4">
-              {/* 아이 기본 정보 내 좌/우 영역 나누기 위한 컨테이너 */}
-              <div className="flex items-center justify-between w-full gap-3 md:gap-6">
-                {/* 좌측: 프로필 이미지 */}
-
-                <div className="flex-shrink-0 w-32 h-32 sm:w-36 sm:h-36 md:w-44 md:h-44 relative overflow-hidden rounded-[13px]">
-                  <Image
-                    src={child.profile || DEFAULT_PROFILE_IMAGE_URL}
-                    width={176}
-                    height={176}
-                    alt="profile Image"
-                    className="absolute w-full h-full object-cover"
-                  />
-                </div>
-
-                {/* 우측: 기본 정보 & 수정하기 버튼 */}
-                <div className="flex flex-col justify-start items-start w-full md:w-[calc(100%-176px)] gap-0 lg:gap-3">
-                  {/* 이름 & 생년월일 시작 */}
-                  <div className="flex flex-col justify-between items-start w-full pt-2  gap-2 lg:gap-3 max-lg:pt-0">
-                    {/*  아이 이름 영역 */}
-                    <p className="text-base font-bold text-center text-neutral-900">{child.name}</p>
-
-                    {/* 아이 생일 영역 */}
-                    <div className="flex justify-center items-center gap-1">
-                      {/* 케이크 아이콘 */}
-                      <Image src={CakeIcon} alt="케이크 아이콘" />
-
-                      {/* 아이 생일 */}
-                      <p className="text-text-s font-medium text-center text-gray-400">{child.birth}</p>
-                    </div>
-                  </div>
-                  {/* 이름 & 생년월일 끝 */}
-
-                  {/* 아이 특이사항 */}
-                  <p className="w-full h-[50px] text-text-s font-medium text-left text-gray-600 overflow-y-auto">
-                    {child.notes && <span>{child.notes}</span>}
-                  </p>
-
-                  {/* 수정하기 */}
-                  <Link key={child.id} href={`/child/${child.id}/childinfo`} className="w-full">
-                    <div className="flex justify-center items-center h-8 gap-[8px] px-4 py-2 rounded-[7px] bg-gray-30">
-                      <button className="text-text-s font-medium text-center text-gray-700">수정하기</button>
-                    </div>
-                  </Link>
-                </div>
-              </div>
-            </div>
-          </div>
+          <ChildBasicInfo child={child} />
           {/* 아이 기본 정보 컨테이너 끝 */}
 
           {/* 우리아이 접종 내역 전체 컨테이너 */}
-          <div className="w-full flex justify-center items-start flex-grow h-52 gap-[288px] p-4 rounded-2xl bg-primary-300">
-            <div className="flex flex-col justify-between items-end self-stretch flex-grow">
-              <div className="flex flex-col justify-start items-start self-stretch flex-grow gap-3">
-                {/* 우리 아이 접종 내역 타이틀 컨테이너 */}
-                <div className="flex justify-center items-center self-stretch flex-grow-0 flex-shrink-0 relative gap-4">
-                  <div className="w-10 h-10 bg-white rounded-full flex justify-center items-center">
-                    <Image src={InjectorIcon} alt="주사기 아이콘" />
-                  </div>
-
-                  <div className="flex flex-col justify-center items-start flex-grow relative gap-1">
-                    <p className="flex-grow-0 flex-shrink-0 text-title-m font-bold text-left text-white">
-                      우리 아이 접종 체크리스트
-                    </p>
-                  </div>
-                </div>
-                {/* 우리 아이 접종 내역 타이틀 컨테이너 끝*/}
-
-                {/* 필수 접종 & 선택 접종 디스플레이 컨테이너 */}
-                <div className="flex flex-col justify-start items-start self-stretch flex-grow gap-2">
-                  {/* 필수 접종 & 선택 접종 */}
-                  <div className="flex justify-start items-start self-stretch flex-grow">
-                    <div className="flex justify-center items-center self-stretch flex-grow relative gap-4 p-4 rounded-xl bg-white">
-                      <div className="flex flex-col justify-center items-center self-stretch flex-grow relative gap-4">
-                        <p className="flex-grow-0 flex-shrink-0 text-title-xs text-left text-gray-600">필수 접종</p>
-                        <p className="flex-grow-0 flex-shrink-0 text-title-m font-semibold text-left text-[#ff7664]">
-                          {requiredVaccinesCount} / {totalRequiredVaccines}개
-                        </p>
-                      </div>
-                      <div style={{ width: 0, height: 60, border: "1px #C2D9FF solid" }}></div>
-                      <div className="flex flex-col justify-center items-center self-stretch flex-grow relative gap-4">
-                        <p className="flex-grow-0 flex-shrink-0 text-title-xs text-left text-gray-600">선택 접종</p>
-                        <p className="flex-grow-0 flex-shrink-0 text-title-m font-semibold text-left text-[#5ebe15]">
-                          {optionalVaccinesCount} / {totalOptionalVaccines}개
-                        </p>
-                      </div>
-                    </div>
-                  </div>
-
-                  {/* 자세히 보기 컨테이너 */}
-                  <div className="flex justify-end items-center self-stretch flex-grow-0 flex-shrink-0 relative gap-1">
-                    <Link href={`/child/${child.id}`}>
-                      <div className="flex flex-col justify-center items-start flex-grow-0 flex-shrink-0 relative gap-1">
-                        <p className="flex-grow-0 flex-shrink-0 text-text-l text-left text-white">자세히 보기</p>
-                      </div>
-                    </Link>
-                    <Image src={RightArrowIcon} alt="오른쪽 화살표 아이콘" />
-                  </div>
-                  {/* 자세히 보기 컨테이너 끝 */}
-                </div>
-                {/* 필수 접종 & 선택 접종 디스플레이 컨테이너 끝*/}
-              </div>
-            </div>
-          </div>
+          <ChildCardVaccinationInfo
+            child={child}
+            requiredVaccinesCount={requiredVaccinesCount}
+            totalRequiredVaccines={totalRequiredVaccines}
+            optionalVaccinesCount={optionalVaccinesCount}
+            totalOptionalVaccines={totalOptionalVaccines}
+          />
           {/* 우리아이 접종 내역 전체 컨테이너 끝*/}
         </div>
         {/* 아이 기본 정보 & 우리 아이 접종 내역 컨테이너 끝*/}
